requirements.txt
pyaudio>=0.2.14
numpy>=1.24.0
requests>=2.31.0
google-genai>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
customtkinter>=5.2.0
